Revert "Add build tags to signal package (#1119)"
This reverts commit 27e432a5b1acb8b8d49629950d84cad11e0a75c6.
