+++ signal/ios.go
@@ -1,0 +1,12 @@
+// +build darwin,cgo
+
+package signal
+
+/*
+#cgo CFLAGS: -x objective-c
+#cgo LDFLAGS: -framework Foundation
+#include <stddef.h>
+#include <stdbool.h>
+extern bool StatusServiceSignalEvent( const char *jsonEvent );
+*/
+import "C"
